Renamed index.js to /utils/container.util.js, uncommented .vscode in gitignores
