+++ .gitignore
@@ -40,7 +40,7 @@ build/Release
 # Dependency directories
 node_modules/
 jspm_packages/
-#.vscode/
+.vscode/
 
 # TypeScript v1 declaration files
 typings/
